[SPLASH] Play "SEGA" choir via PCM, using data sampled from Sonic 3's own "SEGA" digitized sample and exported as a unicode string
The unicode string replaces the pcm_data._sega_choir symbol only after minify to avoid issues with p8tool listrawlua not liking unicode characters like 'x7f' and changing others to underscores '_'.

Thanks to IMLXH (also carlc27843 and czarlo) for the whole conversion procedure (using ffmpeg + p8scii-encoder) as well as the play/load_pcm methods.

Note: backward playing is not used, but was kept in code. In fact, playing "SEGA" backward surprisingly sounds like "SAGE" but with some audio glitches, so we'll still need a handmade version of the "SAGE" choir for the final version.